requirements.txt
django==4.2.3
djangorestframework==3.14.0
psycopg[c]==3.1.9
django-countries==7.5.1 
